restructure with apps folder
