+++ nest-cli.json
@@ -31,7 +31,7 @@
         "kc-db-microservice": {
             "type": "application",
             "root": "src",
-            "entryFile": "src/kc-db-microservice/main",
+            "entryFile": "src/apps/kc-db-microservice/main",
             "sourceRoot": "src",
             "compilerOptions": {
                 "tsConfigPath": "tsconfig.build.json"
